+++ src/passes/CMakeLists.txt
@@ -43,6 +43,7 @@ set(passes_SOURCES
   GlobalStructInference.cpp
   GlobalTypeOptimization.cpp
   GUFA.cpp
+  HashStringifyWalker.cpp
   Heap2Local.cpp
   I64ToI32Lowering.cpp
   Inlining.cpp
